chore: split Navigator
